adjust error message to cleanup code
