+++ lnx-engine/search-index/src/lib.rs
@@ -14,6 +14,7 @@ mod stop_words;
 mod storage;
 pub mod structures;
 mod writer;
+mod schema;
 
 pub use helpers::cr32_hash;
 pub use index::Index;
